Added "show" function
I use Looper for YouTube, and noticed that the button for this would not appear if I clicked on a new video link when used in conjunction with this plugin. So I figured I would add a "show" function in addition to the "hide" and "delete" functions. I added the buttons that appear below YouTube videos ("Add To" / "Share" / "More")